# src/main/resources/api/openapi.yaml
openapi: 3.0.3
info:
  title: Sample Ecommerce App
  description: >
    'This is a ***sample ecommerce app API***.  You can find out more about Swagger at [swagger.io](http://swagger.io).
    Description supports markdown markup. For example, you can use the `inline code` using back ticks.'
  termsOfService: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  contact:
    email: support@packtpub.com
  license:
    name: MIT
    url: https://github.com/PacktPublishing/Modern-API-Development-with-Spring-and-Spring-Boot/blob/master/LICENSE
  version: 1.0.0
...
servers:
  - url: /api/v1  # Relative path which maps to localhost:8080 on dev
    description: Current server
...
paths:
...
  /customers:
    get:
      tags:
        - admin
      summary: Retrieves all customers in the system
      description: Retrieves all customers, or empty collection if no use found. Typically an admin-only endpoint.
      operationId: getAllCustomers
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Customer'
    post:
      tags:
        - customer
      summary: Creates a new customer
      description: Creates a new customer. Throws error if address already exists.
      operationId: createCustomer
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/CustomerReq'
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerReq'
      responses:
        201: # 201 for resource creation
          description: The customer resource was successfully created.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
  /customers/{id}:
    get:
      tags:
        - customer
      summary: Retrieves a customer
      description: Retrieves a customers identifiable by given ID
      operationId: getCustomerById
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
    put:
      tags:
        - customer
      summary: Updates a customer
      description: Updates a customer.
      operationId: updateCustomer
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: true
        description: Customer object (excluding ID which is provided on the path)
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/CustomerReq'
          application/json:
            schema:
              $ref: '#/components/schemas/CustomerReq'
      responses:
        200:
          description: Customer updated successfully
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Customer'
            application/json:
              schema:
                $ref: '#/components/schemas/Customer'
        404:
          description: Given customer ID doesn't exist
          content: { }
    delete:
      tags:
        - customer
      summary: Deletes the customer
      description: Deletes the customer identifiable by given ID
      operationId: deleteCustomerById
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        202:
          description: Request accepted, returns this status even if customer does not exist
          content: {}
  /customers/{id}/addresses:
    get:
      tags:
        - address
      summary: Retrieves the addresses for a specific customer
      description: Retrieves a collection of addresses associated with the given customer ID.
      operationId: getCustomerAddresses
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch. Returns a collection of addresses.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Address'
            application/json:
              schema:
                type: array # Ensure the JSON schema is an array, matching the XML type
                items:
                  $ref: '#/components/schemas/Address'
    post:
      tags:
        - address
      summary: Creates a new customer addresses
      description: Creates a new customer addresses. Does nothing if address already exists.
      operationId: createAddress
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/AddressReq'
          application/json:
            schema:
              $ref: '#/components/schemas/AddressReq'
      responses:
        201: # 201 for resource creation
          description: The address resource was successfully created.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
  /customers/{id}/cards:
    get:
      tags:
        - card
      summary: Retrieves the cards for a specific customer
      description: Retrieves a collection of cards associated with the given customer ID.
      operationId: getCustomerCards
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch. Returns a collection of cards.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Card'
            application/json:
              schema:
                type: array # Ensure the JSON schema is an array, matching the XML type
                items:
                  $ref: '#/components/schemas/Card'
    post:
      tags:
        - card
      summary: Creates a new customer card
      description: Creates a new customer card. Does nothing if card already exists.
      operationId: registerCard
      parameters:
        - name: id
          in: path
          required: true
          description: Customer Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      requestBody:
        required: true
        content:
          application/xml:
            schema:
              $ref: '#/components/schemas/CardReq'
          application/json:
            schema:
              $ref: '#/components/schemas/CardReq'
      responses:
        201: # 201 for resource creation
          description: The card resource was successfully created.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Card'
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
...
  /addresses/{id}:
    get:
      tags:
        - address
      summary: Retrieves address
      description: Retrieves a customer address based on given address ID.
      operationId: getAddressById
      parameters:
        - name: id
          in: path
          required: true
          description: address Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Address'
            application/json:
              schema:
                $ref: '#/components/schemas/Address'
    delete:
      tags:
        - address
      summary: Deletes customer's address
      description: Deletes customer's address based on given address ID.
      operationId: deleteAddressById
      parameters:
        - name: id
          in: path
          required: true
          description: address Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: {}
...
  /cards/{id}:
    get:
      tags:
        - card
      summary: Retrieves a card
      description: Retrieves a card based on given card ID.
      operationId: getCardById
      parameters:
        - name: id
          in: path
          required: true
          description: card Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        200:
          description: For successful fetch.
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Card'
            application/json:
              schema:
                $ref: '#/components/schemas/Card'
    delete:
      tags:
        - card
      summary: Deletes card's address
      description: Deletes card's address based on given card ID.
      operationId: deleteCardById
      parameters:
        - name: id
          in: path
          required: true
          description: card Identifier
          schema:
            type: string
            format: uuid
            pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      responses:
        202:
          description: Accepts the deletion request and perform deletion. If ID does not exist, does nothing.
          content: {}
  /orders:
    get:
      tags:
        - admin
      summary: Retrieves all orders in the system
      description: Retrieves all orders. Typically an admin-only endpoint.
      operationId: getAllOrders
      responses:
        200:
          description: If order exists.
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'
        404:
          description: Order doesn't exist for given customer.
          content: { }
...
components:
  schemas:
    UUID:
      type: string
      description: A standard unique identifier
      format: uuid
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$'
      example: 00000000-0000-0000-0000-000000000001
    BigDecimal:
      type: string
      description: High-precision decimal number. Transmitted as a string to preserve precision.
      example: "123.45"
      # Pattern for validation, ensures the string contains only a number, an optional minus sign at the beginning, and an optional decimal part:
      pattern: '^-?\d+(\.\d+)?$'
...
    CustomerBase:
      type: object
      description: Base properties of a customer object
      required:
        - username
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 40
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        status:
          type: string
      xml:
        name: CustomerBase
    CustomerReq:
      type: object
      description: Request object for a new customer.
      allOf:
        - $ref: '#/components/schemas/CustomerBase'
      xml:
        name: CustomerReq
    Customer:
      type: object
      description: "Represents a complete customer resource, including its ID"
      allOf:
        - $ref: '#/components/schemas/CustomerBase'
      required:
        - id
      properties:
        id:
          description: The id of the customer
          $ref: '#/components/schemas/UUID'
      xml:
        name: Customer
    AddressBase:
      type: object
      description: Base properties for an address
      required:
        - street
        - city
        - zipcode
      properties:
        number:
          type: string
          description: house or flat number
          maxLength: 20
        residency:
          type: string
          description: Society or building name (optional)
        street:
          type: string
          description: street name
          minLength: 1
          maxLength: 80
        city:
          type: string
          description: city name
          minLength: 1
          maxLength: 40
        state:
          type: string
          description: state name
          maxLength: 40
        country:
          type: string
          description: country name
          maxLength: 40
        zipcode:
          type: string
          description: postal code
          minLength: 1
          maxLength: 20
      xml:
        name: AddressBase
    AddressReq:
      type: object
      description: Request object for a new address
      allOf:
        - $ref: '#/components/schemas/AddressBase'
      xml:
        name: AddressReq
    Address:
      type: object
      description: "Represents a complete address resource, including its ID and associated Customer ID."
      allOf:
        - $ref: '#/components/schemas/AddressBase'
      required:
        - id
        - customerId
      properties:
        id:
          description: Address identifier
          $ref: '#/components/schemas/UUID'
        customerId:
          description: The customer ID associated with this address.
          $ref: '#/components/schemas/UUID'
      xml:
        name: Address
    CardDetailsBase:
      type: object
      description: Base properties for a credit/debit card.
      required:
        - cardNumber
        - expires
        - cvv
      properties:
        cardNumber:
          type: string
          description: Card Number (The full number should only be sent in the request, not returned)
        expires:
          type: string
          description: Expiration date (e.g., MM/YY or MM/YYYY)
          example: 01/29
          pattern: '^(0[1-9]|1[0-2])\/\d{2}(\d{2})?$' # Example: MM/YY or MM/YYYY
        cvv:
          type: string
          description: CVV code (Only sent in the request, never stored or returned)
          example: 123
          pattern: '^\d{3,4}$' # 3 or 4 digits
          writeOnly: true
      xml:
        name: CardDetailsBase
    CardReq:
      type: object
      description: Request object for new card registration.
      allOf:
        - $ref: '#/components/schemas/CardDetailsBase'
      xml:
        name: CardReq
    Card:
      type: object
      description: Represents a registered Card resource, with masked details.
      allOf:
          - $ref: '#/components/schemas/CardDetailsBase'
      required:
        - id
        - customerId
        - cardNumber
        - expires
      properties:
        id:
          description: Card identifier
          $ref: '#/components/schemas/UUID'
        customerId:
          description: Id of customer owning the card
          $ref: '#/components/schemas/UUID'
        cardNumber:
          type: string
          description: Card Number (Masked for security)
          example: '************1234'
        expires:
          type: string
          description: Expiration date (MM/YY or MM/YYYY)
        cvv:
          type: string
          description: CVV code (Masked response for security)
          readOnly: true
          example: '***'
      xml:
        name: Card
...
    Item:
      type: object
      description: Item in shopping cart
      required:
        - productId
        - quantity
        - unitPrice
      properties:
        productId:
          description: Product Identifier for this item
          $ref: '#/components/schemas/UUID'
        quantity:
          type: integer
          description: The item quantity
          format: int32
          minimum: 0
        unitPrice:
          description: The item's price per unit
          allOf:
            - $ref: '#/components/schemas/BigDecimal'
      xml:
        name: Item
...
    OrderBase:
      type: object
      description: Base properties for an order, including the items list.
      properties:
        items:
          type: array
          description: Collection of items for the order.
          items:
            $ref: '#/components/schemas/Item'
      xml:
        name: OrderBase
...
    Order:
      type: object
      description: Represents an order
      allOf:
        - $ref: '#/components/schemas/OrderBase'
      required:
        - id
        - date
        - total
        - status
        - payment
        - shipment
      properties:
        id:
          description: Order identifier
          $ref: '#/components/schemas/UUID'
        customer:
          $ref: '#/components/schemas/Customer'
        address:
          $ref: '#/components/schemas/Address'
        card:
          $ref: '#/components/schemas/Card'
        date:
          type: string
          description: Order's data and time details
          format: date-time
        total:
          description: Order total
          allOf:
            - $ref: '#/components/schemas/BigDecimal'
        payment:
          $ref: '#/components/schemas/Payment'
        shipment:
          $ref: '#/components/schemas/Shipment'
        status:
          type: string
          description: Order Status
          enum:
            - CREATED
            - PAID
            - SHIPPED
            - PAYMENT_FAILED
            - SHIPMENT_FAILED
            - COMPLETED
      xml:
        name: Order
...
    Payment:
      type: object
      description: Object corresponding to a payment.
      properties:
        id:
          description: Payment identifier
          $ref: '#/components/schemas/UUID'
        orderId:
          description: Order Identifier belonging to this payment request
          $ref: '#/components/schemas/UUID'
        authorized:
          type: boolean
          description: Flag that specifies whether payment is authorized or not
        message:
          type: string
          description: Approval or rejection message
      xml:
        name: Payment
...
    Shipment:
      type: object
      description: Object corresponding to a shipment.
      properties:
        id:
          description: Shipment Identifier
          $ref: '#/components/schemas/UUID'
        orderId:
          description: Order Identifier
          $ref: '#/components/schemas/UUID'
        carrier:
          type: string
          description: Shipping Carrier
        trackingNumber:
          type: string
          description: Shipping Tracking Number
        estDeliveryDate:
          type: string
          description: Estimated Delivery Date
          format: date
      xml:
        name: Shipment